+++ Snag/App/SnagApp.swift
@@ -1,0 +1,8 @@
+import SwiftUI
+
+@main
+struct SnagApp: App {
+    var body: some Scene {
+        WindowGroup { RootView() }
+    }
+}